paddlepaddle/requirements.txt
pycocotools
torch
torchvision
numpy
argparse
pyyaml
paddlepaddle
paddle2onnx
opencv-python
tqdm
pillow
scipy
six
shapely
pyclipper
imageio
polygon3
onnx==1.11.0
protobuf==3.20.0
h5py
onnxoptimizer
rich
addict
